% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longreadvqs-package.R
\docType{package}
\name{longreadvqs-package}
\alias{longreadvqs}
\alias{longreadvqs-package}
\title{longreadvqs: Viral Quasispecies Comparison from Long-Read Sequencing Data}
\description{
Performs variety of viral quasispecies diversity analyses [see Pamornchainavakul et al. (2026) \doi{10.1371/journal.pcbi.1014208}] based on long-read sequence alignment. Main functions include 1) sequencing error and other noise minimization and read sampling, 2) Single nucleotide variant (SNV) profiles comparison, and 3) viral quasispecies profiles comparison and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NakarinP/longreadvqs}
  \item Report bugs at \url{https://github.com/NakarinP/longreadvqs/issues}
}

}
\author{
\strong{Maintainer}: Nakarin Pamornchainavakul \email{pamornakarin@gmail.com} (\href{https://orcid.org/0000-0003-0378-0316}{ORCID})

}
\keyword{internal}
